fix: fix basename path on github deployment
